Ensure order of repeated values in the `set`
Instead of using a `multiset`, switch to a `set`. To ensure that repeat
values can be handled, use a `pair` to tie each value to a `size_t`
number that indicates how many values have been seen thus far. Not only
does the `size_t` value break ties between duplicate values (and ensure
each `pair` is unique), it additionally ensures that repeat values
maintain their order when inserted into the `set`. This property is
relied on by the rank filter algorithm here to handle special insertion
cases. While C++11's `multiset` and Boost's `multiset` have this
property, we are not sure of the latter and cannot find documentation to
back it up. Thus creating `pair`s that avoids any need to rely on
`multiset`s that may or may not maintain ordering of duplicate values.